+++ packages/components/src/Checkbox/CheckboxIcon.tsx
@@ -5,13 +5,12 @@ import { Box } from '../Box';
 
 // Checkbox Icon
 // ---------------
-export type CheckboxIconProps = {
+export interface CheckboxIconProps {
   variant?: string;
   checked?: boolean;
   disabled?: boolean;
   error?: boolean;
-  children?: never;
-};
+}
 
 export const CheckboxIcon: React.FC<CheckboxIconProps> = ({
   variant = '',